auth and user services divide
