+++ src/main/java/ru/vldf/sportsportal/controller/MatchesController.java
@@ -4,20 +4,20 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.GetMapping;
-import ru.vldf.sportsportal.service.UserService;
+import ru.vldf.sportsportal.service.AuthService;
 
 @Controller
 public class MatchesController {
-    private UserService userService;
+    private AuthService authService;
 
     @Autowired
-    public void setUserService(UserService userService) {
-        this.userService = userService;
+    public void setAuthService(AuthService authService) {
+        this.authService = authService;
     }
 
     @GetMapping(value = {"/matches"})
     public String coverPage(ModelMap map) {
-        map.addAttribute("username", userService.getAuthUsername());
+        map.addAttribute("username", authService.getAuthUsername());
 
         return "matches/cover";
     }
